+++ tests/testthat/test-remove.zero.brlen.R
@@ -37,3 +37,29 @@ test_that("remove.zero.brlen works", {
     message <- capture_output(remove.zero.brlen(tree, verbose = TRUE))
     expect_equal(message, "Changing 5 branch lengths:.....Done.")
 })
+
+test_that("remove.zero.brlen works with multiPhylo", {
+    ## Generating some trees
+    trees <- rmtree(10, 10)
+    trees[[1]]$edge.length[2] <- 0
+    expect_true(any(unlist(lapply(trees, function(x) return(x$edge.length))) == 0))
+    ## Removing all 0s
+    trees2 <- remove.zero.brlen(trees)
+    expect_is(trees2, "multiPhylo")
+    expect_false(any(unlist(lapply(trees2, function(x) return(x$edge.length))) == 0))
+})
+
+test_that("remove.zero.brlen also removes negative brlen", {
+    ## Generating a tree with negative branch lengths
+    set.seed(3)
+    tree <- rtree(10)
+    tree_neg <- chronoMPL(tree)
+    # plot(tree_neg, main = "Negative branch lengths")
+    # plot(tree_pos, main = "Positive branch lengths")
+    expect_true(any(tree_neg$edge.length < 0))
+    expect_true(is.ultrametric(tree_neg))
+    ## Removing negative branch length
+    tree_pos <- remove.zero.brlen(tree_neg)
+    expect_true(all(tree_pos$edge.length > 0))
+    expect_true(is.ultrametric(tree_pos))
+})